+++ src/project/contract/llvm_ir.rs
@@ -3,8 +3,8 @@
 //!
 
 use crate::build::contract::Contract as ContractBuild;
+use crate::message_type::MessageType;
 use crate::project::contract::metadata::Metadata as ContractMetadata;
-use crate::warning_type::WarningType;
 
 ///
 /// The LLVM IR contract.
@@ -38,7 +38,7 @@ impl Contract {
         optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
         llvm_options: Vec<String>,
         output_assembly: bool,
-        _suppressed_warnings: Vec<WarningType>,
+        _suppressed_messages: Vec<MessageType>,
         debug_config: Option<era_compiler_llvm_context::DebugConfig>,
     ) -> anyhow::Result<ContractBuild> {
         let llvm = inkwell::context::Context::create();