+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 
 - The vyper v0.4.0 support
 - The `--enable-decimals` parameter
+- The support for compiling multiple files in Yul, LLVM IR, and EraVM assembly modes
 
 ### Changed
 
 - Assembly is now only returned if requested with `--output-assembly`
+- Renamed the `zkasm` flag to `eravm-assembly`
 - Updated to Rust v1.79.0
 
 ### Fixed